Extract base dev AI tools
